requirements.txt
python_ex @ git+https://github.com/SEOULTECH-AIS-gyounghun6612/AIS_python_ex.git
einops
torchsummary
tensorboard
torchvision